refactor: JSONFileManager를 Repository와 병합
