--- src/main/java/team/marco/vouchermanagementsystem/util/FileManager.java
@@ -1,11 +1,0 @@
-package team.marco.vouchermanagementsystem.util;
-
-import java.util.List;
-
-public interface FileManager<T> {
-    List<T> load();
-
-    void save(T data);
-
-    void close();
-}
