+++ src/main/resources/sample_data.json
@@ -1,1 +1,1 @@
-[{"type":"FIXED","id":"461b8847-578d-43c7-8472-d9374bbac41a","data":1000000},{"type":"PERCENT","id":"7bd6cc26-5c3b-4579-ad3b-834f2b6f075c","data":50},{"type":"FIXED","id":"c0f9c0f3-b974-41e8-a517-df7c3b8cd08b","data":100},{"type":"PERCENT","id":"75145684-5602-40a2-b279-096d94dc6063","data":22},{"type":"FIXED","id":"da9aa8de-162f-46c7-b23b-2e37ee0569aa","data":7777},{"type":"PERCENT","id":"4d3cb1bc-02f5-4d67-b274-60526d87fac1","data":90}]
+[{"id":"861dbf02-3989-467c-b0a5-7c8b56640688","type":"FIXED","data":1000,"info":"1,000원 할인 쿠폰"},{"id":"9f749266-b875-431b-b728-d17ef34ca1ee","type":"PERCENT","data":10,"info":"10% 할인 쿠폰"}]
